refactor(status): address PR review feedback
Move subtypes and timers into individual handlers so they can
control their own state. StatusHandlers becomes a Singleton with
instance methods. Handlers own their subtypes, timers, and display
logic.

- Remove thread lock (Python dicts are thread-safe)
- Use typing.Type instead of quoted type hints
- Integer timestamps in DB, not float
- Remove redundant migration (table created by DB_SCHEMA)
- Remove verbose logging on clear
- Add get_expiry to StatusHandler ABC with default None
- Add restore method for loading from DB with remaining time
- Rename is_active to problem_reported
- Move subtype labels to frontend JavaScript
- Status page uses list format with clear button per type
- Badge uses yellow square to match other *arr apps
- Library import banner centered below action buttons
- API DELETE accepts optional type param for per-type clear
- Add unit test for handler registration coverage

+++ backend/internals/db.py
@@ -531,8 +531,8 @@ CREATE TABLE IF NOT EXISTS remote_mappings(
 CREATE TABLE IF NOT EXISTS status(
     status_type VARCHAR(100) NOT NULL,
     subtype VARCHAR(100) NOT NULL,
-    timestamp REAL NOT NULL,
-    expires_at REAL,
+    timestamp INTEGER NOT NULL,
+    expires_at INTEGER,
     PRIMARY KEY (status_type, subtype)
 );
 """
